Added option to set the title of the application. Fixed some bugs and made sure it works in Ubuntu/Chrome, Ubuntu/Opera and Windows/IE8.

+++ app/models/Upload.java
@@ -124,7 +124,9 @@ public class Upload extends Model {
     
     @Override
     public void delete() {
-    	getFile().delete();
+    	File file = getFile();
+    	if (file.exists())
+    		getFile().delete();
     	super.delete();
     }
 }
